Support coarray data of which is more than 2GB, and transfer more than 16MB data on the K computer

+++ XcodeML-Exc-Tools/src/exc/xcalablemp/XMPrewriteExpr.java
@@ -1397,7 +1397,7 @@ public class XMPrewriteExpr {
     for(int i=0; i<coarray.getVarDim(); i++){
       Xobject tmp = null;
       for(int j=coarray.getVarDim()-1; j>i; j--){
-        int size = coarray.getSizeAt(j);
+	  int size = (int)coarray.getSizeAt(j);
         if(tmp == null){
           tmp = Xcons.Int(Xcode.INT_CONSTANT, size);
         } else{
